complete the update
