+++ test/REQUIRE
@@ -1,5 +1,2 @@
-ParameterizedFunctions
 DiffEqDevTools
 DiffEqProblemLibrary
-Knet
-DiffEqBase
